doc-analyser/requirements.txt
omegaconf
pytorch-lightning
colorlog
rootutils